MIWM-368: Change payments coordination
