+++ src/main/java/miwm/job4me/services/event/JobFairService.java
@@ -1,9 +1,11 @@
 package miwm.job4me.services.event;
 
 import miwm.job4me.model.event.JobFair;
+import miwm.job4me.model.payment.Payment;
 import miwm.job4me.services.BaseDtoService;
 import miwm.job4me.web.model.event.JobFairDto;
 import miwm.job4me.web.model.listDisplay.ListDisplayDto;
+import miwm.job4me.web.model.payment.PaymentCheckout;
 import org.springframework.data.domain.Page;
 
 public interface JobFairService extends BaseDtoService<JobFair, JobFairDto, Long> {
@@ -18,7 +20,7 @@ public interface JobFairService extends BaseDtoService<JobFair, JobFairDto, Long
 
     Page<ListDisplayDto> findAllOfSignedInOrganizerByFiltersListDisplay(int page, int size, String order, Boolean showUpcoming, String address, Boolean isPaid);
 
-    JobFairDto saveDto(JobFairDto jobFairDto);
+    JobFairDto saveDto(JobFairDto jobFairDto, Payment payment);
 
     boolean existsById(Long id);
 
@@ -31,5 +33,7 @@ public interface JobFairService extends BaseDtoService<JobFair, JobFairDto, Long
     JobFair getJobFairById(Long id);
 
     JobFair getOnlyPaidJobFairById(Long id);
+
+    PaymentCheckout coordinateJobFairPayment(JobFairDto jobFairDto);
 
 }
