+++ src/main/java/miwm/job4me/services/event/JobFairServiceImpl.java
@@ -1,11 +1,14 @@
 package miwm.job4me.services.event;
 
+import com.stripe.model.checkout.Session;
 import miwm.job4me.exceptions.InvalidArgumentException;
 import miwm.job4me.exceptions.NoSuchElementFoundException;
 import miwm.job4me.messages.ExceptionMessages;
 import miwm.job4me.model.event.JobFair;
+import miwm.job4me.model.payment.Payment;
 import miwm.job4me.model.users.Organizer;
 import miwm.job4me.repositories.event.JobFairRepository;
+import miwm.job4me.services.payment.PaymentService;
 import miwm.job4me.services.users.OrganizerService;
 import miwm.job4me.validators.entity.event.JobFairValidator;
 import miwm.job4me.validators.fields.IdValidator;
@@ -14,6 +17,7 @@ import miwm.job4me.web.mappers.event.JobFairMapper;
 import miwm.job4me.web.mappers.listDisplay.ListDisplayMapper;
 import miwm.job4me.web.model.event.JobFairDto;
 import miwm.job4me.web.model.listDisplay.ListDisplayDto;
+import miwm.job4me.web.model.payment.PaymentCheckout;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -33,6 +37,8 @@ public class JobFairServiceImpl implements JobFairService {
     private final IdValidator idValidator;
     private final PaginationValidator paginationValidator;
     private final OrganizerService organizerService;
+
+    private final PaymentService paymentService;
     private final String ENTITY_NAME = "JobFair";
 
     private final Map<String, Sort> orderMap = Map.of(
@@ -41,7 +47,7 @@ public class JobFairServiceImpl implements JobFairService {
             "3", Sort.by(Sort.Direction.DESC, "dateStart")
     );
 
-    public JobFairServiceImpl(JobFairRepository jobFairRepository, JobFairMapper jobFairMapper, ListDisplayMapper listDisplayMapper, JobFairValidator jobFairValidator, IdValidator idValidator, PaginationValidator paginationValidator, OrganizerService organizerService) {
+    public JobFairServiceImpl(JobFairRepository jobFairRepository, JobFairMapper jobFairMapper, ListDisplayMapper listDisplayMapper, JobFairValidator jobFairValidator, IdValidator idValidator, PaginationValidator paginationValidator, OrganizerService organizerService, PaymentService paymentService) {
         this.jobFairRepository = jobFairRepository;
         this.jobFairMapper = jobFairMapper;
         this.listDisplayMapper = listDisplayMapper;
@@ -49,6 +55,7 @@ public class JobFairServiceImpl implements JobFairService {
         this.idValidator = idValidator;
         this.paginationValidator = paginationValidator;
         this.organizerService = organizerService;
+        this.paymentService = paymentService;
     }
 
     @Override
@@ -141,12 +148,14 @@ public class JobFairServiceImpl implements JobFairService {
     }
 
     @Override
-    public JobFairDto saveDto(JobFairDto jobFairDto) {
+    public JobFairDto saveDto(JobFairDto jobFairDto, Payment payment) {
 //        idValidator.validateNoIdForCreate(jobFairDto.getId(), ENTITY_NAME);
         Organizer organizer = organizerService.getAuthOrganizer();
-        jobFairDto.setOrganizerId(organizer.getId());
         jobFairValidator.validateDto(jobFairDto);
-        return jobFairMapper.toDto(jobFairRepository.save(jobFairMapper.toEntity(jobFairDto)));
+        JobFair jobFair = jobFairMapper.toEntity(jobFairDto);
+        jobFair.setOrganizer(organizer);
+        jobFair.setPayment(payment);
+        return jobFairMapper.toDto(jobFairRepository.save(jobFair));
     }
 
     @Override
@@ -196,6 +205,25 @@ public class JobFairServiceImpl implements JobFairService {
         }
 
         return jobFair;
+    }
+
+    @Override
+    public PaymentCheckout coordinateJobFairPayment(JobFairDto jobFairDto) {
+        Organizer organizer = organizerService.getAuthOrganizer();
+        Session session = paymentService.createJobFairPayment(organizer.getEmail());
+
+        Payment payment = Payment.builder()
+                .isPaid(false)
+                .sessionId(session.getId())
+                .creationTimestamp(LocalDateTime.now())
+                .build();
+
+        saveDto(jobFairDto, payment);
+
+        PaymentCheckout paymentCheckout = new PaymentCheckout();
+        paymentCheckout.setUrl(session.getUrl());
+
+        return paymentCheckout;
     }
 
 }
