+++ src/main/java/miwm/job4me/services/payment/PaymentService.java
@@ -2,8 +2,6 @@ package miwm.job4me.services.payment;
 
 import com.stripe.model.checkout.Session;
 import miwm.job4me.model.payment.Payment;
-import miwm.job4me.web.model.event.JobFairDto;
-import miwm.job4me.web.model.payment.PaymentCheckout;
 
 public interface PaymentService {
 
@@ -13,11 +11,9 @@ public interface PaymentService {
 
     Payment getPaymentBySessionId(String sessionId);
 
-    PaymentCheckout coordinateJobFairPayment(JobFairDto jobFairDto);
-
     void handleCheckoutPayment(String payload, String sigHeader);
 
-    Session createJobFairPayment();
+    Session createJobFairPayment(String customerEmail);
 
     Session createPaymentSession(Long productQuantity, Long productPrice, String productName, String successUrl, String cancelUrl, String productImageUrl, String productDescription, String customerEmail);
