+++ src/main/java/miwm/job4me/services/payment/PaymentServiceImpl.java
@@ -10,18 +10,11 @@ import com.stripe.net.Webhook;
 import com.stripe.param.checkout.SessionCreateParams;
 import miwm.job4me.exceptions.PaymentException;
 import miwm.job4me.model.payment.Payment;
-import miwm.job4me.model.users.Organizer;
 import miwm.job4me.repositories.payment.PaymentRepository;
-import miwm.job4me.services.event.JobFairService;
-import miwm.job4me.services.users.OrganizerService;
 import miwm.job4me.validators.entity.payment.PaymentValidator;
 import miwm.job4me.validators.fields.IdValidator;
-import miwm.job4me.web.model.event.JobFairDto;
-import miwm.job4me.web.model.payment.PaymentCheckout;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
 
 import static miwm.job4me.messages.AppMessages.FRONT_HOST_AZURE;
 
@@ -37,15 +30,11 @@ public class PaymentServiceImpl implements PaymentService {
     private final PaymentRepository paymentRepository;
     private final PaymentValidator paymentValidator;
     private final IdValidator idValidator;
-    private final OrganizerService organizerService;
-    private final JobFairService jobFairService;
 
-    public PaymentServiceImpl(PaymentRepository paymentRepository, PaymentValidator paymentValidator, IdValidator idValidator, OrganizerService organizerService, JobFairService jobFairService) {
+    public PaymentServiceImpl(PaymentRepository paymentRepository, PaymentValidator paymentValidator, IdValidator idValidator) {
         this.paymentRepository = paymentRepository;
         this.paymentValidator = paymentValidator;
         this.idValidator = idValidator;
-        this.organizerService = organizerService;
-        this.jobFairService = jobFairService;
     }
 
     @Override
@@ -65,26 +54,6 @@ public class PaymentServiceImpl implements PaymentService {
     public Payment getPaymentBySessionId(String sessionId) {
         idValidator.validateStringId(sessionId, "Session");
         return paymentRepository.findBySessionId(sessionId);
-    }
-
-    @Override
-    public PaymentCheckout coordinateJobFairPayment(JobFairDto jobFairDto) {
-        JobFairDto jobFair = jobFairService.saveDto(jobFairDto);
-        Session session = createJobFairPayment();
-
-        Payment payment = Payment.builder()
-                .isPaid(false)
-                .jobFair(jobFairService.getJobFairById(jobFair.getId()))
-                .sessionId(session.getId())
-                .creationTimestamp(LocalDateTime.now())
-                .build();
-
-        save(payment);
-
-        PaymentCheckout paymentCheckout = new PaymentCheckout();
-        paymentCheckout.setUrl(session.getUrl());
-
-        return paymentCheckout;
     }
 
     @Override
@@ -118,9 +87,7 @@ public class PaymentServiceImpl implements PaymentService {
     }
 
     @Override
-    public Session createJobFairPayment() {
-        Organizer organizer = organizerService.getAuthOrganizer();
-
+    public Session createJobFairPayment(String customerEmail) {
         Long productQuantity = 1L;
         Long productPrice = 200L;
         String productName = "Utworzenie targów pracy";
@@ -128,7 +95,6 @@ public class PaymentServiceImpl implements PaymentService {
         String cancelUrl = FRONT_HOST_AZURE + "/organizer/payment/cancel";
         String productImageUrl = "https://files.stripe.com/links/MDB8YWNjdF8xTzlaYWRJb1RMYU5hVEFqfGZsX2xpdmVfMXpMRkZueXpwc0VQaFdjOXNiU2p3a1Zp00kFAIQX0R";
         String productDescription = "Tworzy targi pracy i umożliwia ich przeprowadzenie.";
-        String customerEmail = organizer.getEmail();
 
         return createPaymentSession(productQuantity, productPrice, productName, successUrl, cancelUrl, productImageUrl, productDescription, customerEmail);
     }
