+++ src/main/java/miwm/job4me/web/controllers/payment/PaymentController.java
@@ -2,8 +2,6 @@ package miwm.job4me.web.controllers.payment;
 
 import io.swagger.v3.oas.annotations.Operation;
 import miwm.job4me.services.payment.PaymentService;
-import miwm.job4me.web.model.event.JobFairDto;
-import miwm.job4me.web.model.payment.PaymentCheckout;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -26,14 +24,6 @@ public class PaymentController {
         paymentService.handleCheckoutPayment(payload, sigHeader);
 
         return new ResponseEntity<>("Webhook received successfully", HttpStatus.OK);
-    }
-
-    @PostMapping("organizer/job-fairs/payment")
-    @Operation(summary = "Create JobFair and payment session", description = "Create JobFair and payment session")
-    public ResponseEntity<PaymentCheckout> payForOrganizerAccount(@RequestBody JobFairDto jobFairDto) {
-        PaymentCheckout paymentCheckout = paymentService.coordinateJobFairPayment(jobFairDto);
-
-        return new ResponseEntity<>(paymentCheckout, HttpStatus.OK);
     }
 
 }
